pyproject.toml
[project]
name = "langagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
